refactor: extract routers/practice.py
Move question/evaluate/submit/bank/history/sessions routes out of
main.py into api/routers/practice.py. Extract _parse_result_json helper
to eliminate the duplicate result_json parsing block in get_session_results
and get_history.
