+++ tests/api/test_api_ui.py
@@ -102,7 +102,7 @@ def test_get_ui_without_query_returns_404_for_missing_context(client: TestClient
 def test_get_question_fragment_returns_200(client: TestClient, mock_retriever: MagicMock) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
     with patch(
-        "learning_tool.api.main.generate_question_gemini",
+        "learning_tool.api.routers.practice.generate_question_gemini",
         new=AsyncMock(return_value=Question(text="What is X?")),
     ):
         response = client.get("/ui/my-context/question?query=topic&session_id=test-session-id")
@@ -116,7 +116,7 @@ def test_get_question_fragment_contains_question_id(
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
     with patch(
-        "learning_tool.api.main.generate_question_gemini",
+        "learning_tool.api.routers.practice.generate_question_gemini",
         new=AsyncMock(return_value=Question(text="What is X?")),
     ):
         response = client.get("/ui/my-context/question?query=topic&session_id=test-session-id")
@@ -129,7 +129,7 @@ def test_get_question_fragment_contains_question_text(
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
     with patch(
-        "learning_tool.api.main.generate_question_gemini",
+        "learning_tool.api.routers.practice.generate_question_gemini",
         new=AsyncMock(return_value=Question(text="What is X?")),
     ):
         response = client.get("/ui/my-context/question?query=topic&session_id=test-session-id")
@@ -142,7 +142,7 @@ def test_get_question_fragment_contains_answer_form(
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
     with patch(
-        "learning_tool.api.main.generate_question_gemini",
+        "learning_tool.api.routers.practice.generate_question_gemini",
         new=AsyncMock(return_value=Question(text="What is X?")),
     ):
         response = client.get("/ui/my-context/question?query=topic&session_id=test-session-id")
@@ -156,7 +156,7 @@ def test_get_question_fragment_form_posts_to_evaluate(
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
     with patch(
-        "learning_tool.api.main.generate_question_gemini",
+        "learning_tool.api.routers.practice.generate_question_gemini",
         new=AsyncMock(return_value=Question(text="What is X?")),
     ):
         response = client.get("/ui/my-context/question?query=topic&session_id=test-session-id")
@@ -169,7 +169,7 @@ def test_get_question_fragment_passes_query_to_form(
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
     with patch(
-        "learning_tool.api.main.generate_question_gemini",
+        "learning_tool.api.routers.practice.generate_question_gemini",
         new=AsyncMock(return_value=Question(text="What is X?")),
     ):
         response = client.get("/ui/my-context/question?query=topic&session_id=test-session-id")
@@ -180,7 +180,9 @@ def test_get_question_fragment_passes_query_to_form(
 
 def test_post_evaluate_fragment_returns_200(client: TestClient, mock_retriever: MagicMock) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
-    with patch("learning_tool.api.main.evaluate_answer", new=AsyncMock(return_value=EVALUATION)):
+    with patch(
+        "learning_tool.api.routers.practice.evaluate_answer", new=AsyncMock(return_value=EVALUATION)
+    ):
         response = client.post(
             "/ui/my-context/evaluate",
             data={
@@ -199,7 +201,9 @@ def test_post_evaluate_fragment_contains_score(
     client: TestClient, mock_retriever: MagicMock
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
-    with patch("learning_tool.api.main.evaluate_answer", new=AsyncMock(return_value=EVALUATION)):
+    with patch(
+        "learning_tool.api.routers.practice.evaluate_answer", new=AsyncMock(return_value=EVALUATION)
+    ):
         response = client.post(
             "/ui/my-context/evaluate",
             data={
@@ -219,7 +223,9 @@ def test_post_evaluate_fragment_uses_query_for_retrieval(
     client: TestClient, mock_retriever: MagicMock
 ) -> None:
     mock_retriever.retrieve.return_value = [("chunk", 0.9)]
-    with patch("learning_tool.api.main.evaluate_answer", new=AsyncMock(return_value=EVALUATION)):
+    with patch(
+        "learning_tool.api.routers.practice.evaluate_answer", new=AsyncMock(return_value=EVALUATION)
+    ):
         client.post(
             "/ui/my-context/evaluate",
             data={
